correct size_t hhh thing it is ok hh blida tgarbi3it to5mama

+++ 0-bubble_sort.c
@@ -10,24 +10,24 @@
  */
 void bubble_sort(int *array, size_t size)
 {
-    size_t temp = 0;
-    size_t i = 0, j = 0;
+	int temp = 0;
+	size_t i = 0, j = 0;
 
-    if (array == NULL)
-    {
-        return;
-    }
-    for (i = 0; i < size - 1; i++)
-    {
-        for (j = 0; j < size - i; j++)
-        {
-            if (array[j] > array[j + 1])
-            {
-                temp = array[j];
-                array[j] = array[j + 1];
-                array[j + 1] = temp;
-                print_array(array, size);
-            }
-        }
-    }
+	if (array == NULL)
+	{
+		return;
+	}
+	for (i = 0; i < size - 1; i++)
+	{
+		for (j = 0; j < size - i; j++)
+		{
+			if (array[j] > array[j + 1])
+			{
+				temp = array[j];
+				array[j] = array[j + 1];
+				array[j + 1] = temp;
+				print_array(array, size);
+			}
+		}
+	}
 }
